Split ElasticSearchOrganizerSearchServiceTest into ES5 and ES8 variants
